requirements.txt
fastapi
httpx
redis
uvicorn
numpy
aioredis
openai
pydantic
beautifulsoup4
requests
python-jose
prometheus-client
pandas
langchain
faiss-cpu
transformers
torch
sentence_transformers
psutil
passlib
python-dotenv
bcrypt
streamlit